Renamae package to snapshot_generator
