--- data_api/snapshot_convertor/src/main/scala/uk/ac/wellcome/platform/snapshot_convertor/models/CompletedSnapshotJob.scala
@@ -1,8 +1,0 @@
-package uk.ac.wellcome.platform.snapshot_convertor.models
-
-import akka.http.scaladsl.model.Uri
-
-case class CompletedSnapshotJob(
-                                 snapshotJob: SnapshotJob,
-                                 targetLocation: Uri
-)
