--- data_api/snapshot_convertor/src/main/scala/uk/ac/wellcome/platform/snapshot_convertor/models/SnapshotJob.scala
@@ -1,7 +1,0 @@
-package uk.ac.wellcome.platform.snapshot_convertor.models
-
-import uk.ac.wellcome.versions.ApiVersions
-
-case class SnapshotJob(publicBucketName: String,
-                       publicObjectKey: String,
-                       apiVersion: ApiVersions.Value)
