+++ data_api/snapshot_convertor/src/main/scala/uk/ac/wellcome/platform/snapshot_generator/models/CompletedSnapshotJob.scala
@@ -1,0 +1,8 @@
+package uk.ac.wellcome.platform.snapshot_generator.models
+
+import akka.http.scaladsl.model.Uri
+
+case class CompletedSnapshotJob(
+                                 snapshotJob: SnapshotJob,
+                                 targetLocation: Uri
+)
