+++ data_api/snapshot_convertor/src/main/scala/uk/ac/wellcome/platform/snapshot_generator/models/SnapshotJob.scala
@@ -1,0 +1,7 @@
+package uk.ac.wellcome.platform.snapshot_generator.models
+
+import uk.ac.wellcome.versions.ApiVersions
+
+case class SnapshotJob(publicBucketName: String,
+                       publicObjectKey: String,
+                       apiVersion: ApiVersions.Value)
