--- data_api/snapshot_convertor/src/test/scala/uk/ac/wellcome/platform/snapshot_convertor/fixtures/AkkaS3.scala
@@ -1,32 +1,0 @@
-package uk.ac.wellcome.platform.snapshot_convertor.fixtures
-
-import akka.actor.ActorSystem
-import akka.stream.ActorMaterializer
-import akka.stream.alpakka.s3.scaladsl.S3Client
-import com.amazonaws.auth.{AWSStaticCredentialsProvider, BasicAWSCredentials}
-import com.amazonaws.regions.AwsRegionProvider
-import uk.ac.wellcome.platform.snapshot_convertor.modules.AkkaS3ClientModule
-import uk.ac.wellcome.test.fixtures.{S3, TestWith}
-
-trait AkkaS3 extends S3 {
-
-  def withS3AkkaClient[R](
-    actorSystem: ActorSystem,
-    materializer: ActorMaterializer)(testWith: TestWith[S3Client, R]): R = {
-
-    val s3Settings = AkkaS3ClientModule.akkaS3Settings(
-      credentialsProvider = new AWSStaticCredentialsProvider(
-        new BasicAWSCredentials(accessKey, secretKey)
-      ),
-      regionProvider = new AwsRegionProvider {
-        def getRegion: String = regionName
-      },
-      endpointUrl = Some(localS3EndpointUrl)
-    )
-
-    val s3AkkaClient =
-      new S3Client(s3Settings = s3Settings)(actorSystem, materializer)
-
-    testWith(s3AkkaClient)
-  }
-}
